work

+++ main.py
@@ -46,8 +46,8 @@ def mainLoop():
             time.sleep(1)
 
             updateBrightness()
-            lightController.strip.setBrightness(brightness*2.55)
-            lightController.strip2.setBrightness(brightness * 2.55)
+            lightController.strip.setBrightness(int(brightness*2.55))
+            lightController.strip2.setBrightness(int(brightness * 2.55))
 
     finally:
         stopAllThreads()
